refactor: extract the secret lookup and value injection from the injector
The job push needs the same two mechanics the long poll already has:
look the secret references of a job up in the per-store caches, and
replace the placeholders in the job's variables with the values found.
Both lived in JobSecretInjector, wired into the batch bookkeeping of the
long poll (indexes in the batch, growth budget, dropping jobs), which
the push has no use for: it hands out one job at a time.

JobSecretLookup now owns the two mechanics and JobSecretInjector keeps
the batch bookkeeping and delegates. The values travel on the check
result rather than on the lookup: check() returns a CachedSecret per
reference carrying the value its store held, so plaintext secrets live
exactly as long as the result its caller holds and no caller has to
reach a reset to drop them. What is left on the class is the store
registry and the msgpack mapper, which is what its name says it is.

Reading the values at check time is kept. It is what stops a value
evicted between the check and the injection from dropping a job late,
which matters most on the long poll, where the injection runs after the
batch value is already fixed. The cost is that two jobs of one batch
sharing a reference now read the local cache once each rather than once
in total, which is an in-memory read and reads no store.

No behaviour change. The nested Secret and SecretCheckResult records
move along with the mechanics that produce them.
